aider: Fix namespace error in Player.cs by adding using statement for Microsoft.Xna.Framework.
# Aider chat conversation:

USER: make sure all the namespace errors are fixed

ASSISTANT: To fix the namespace errors, we need to ensure that all the classes are using the correct namespaces and that they are correctly referenced in the files where they are used.

From the provided code, it seems like the `Player` class and the `World` class are both under the `EpochsUnbound.Entities` namespace. However, the `Player` class is also using the `Vector2` class from the `Microsoft.Xna.Framework` namespace, and the `World` class is using the `Tile` class, which is not defined in the provided code.

The `Game1` class is using the `AdventurerMode` class from the `EpochsUnbound.GameModes` namespace, and the `GameState` enum defined in the same file. It's also using several classes from the `Microsoft.Xna.Framework` namespace.

Here are the changes needed:

1. Add the `using Microsoft.Xna.Framework;` statement to the `Player.cs` file.
2. Define the `Tile` class or add the correct `using` statement to the `World.cs` file.
3. No changes are needed in the `Game1.cs` file as it seems to be using the correct namespaces.

Here are the edit blocks for these changes:

```csharp
Entities\Player.cs
<<<<<<< HEAD
namespace EpochsUnbound.Entities
{
    public class Player
=======
using Microsoft.Xna.Framework;

namespace EpochsUnbound.Entities
{
    public class Player
>>>>>>> updated
```

For the `World.cs` file, I need to know where the `Tile` class is defined to add the correct `using` statement. Could you please provide this information?