aider: Added basic implementation of the `Tile` class in the `Entities` folder.
# Aider chat conversation:

USER: I'm not sure if it's defined or not

ASSISTANT: If you're unsure whether the `Tile` class is defined or not, you can search for it in your project. If you're using an Integrated Development Environment (IDE) like Visual Studio, you can use the "Find All" feature (usually Ctrl+Shift+F) and search for "class Tile".

If the `Tile` class is not defined, we need to create it. The `Tile` class should at least have `Update` and `Draw` methods since these methods are called in the `World` class.

If you want, I can provide a basic implementation of the `Tile` class. Please let me know where you want to place this class (which directory and namespace).

USER: Lets put it in the Entities folder

ASSISTANT: Alright, let's create a basic `Tile` class in the `Entities` directory. This class will have `Update` and `Draw` methods, which are called in the `World` class.

Here's the edit block for the new `Tile` class:

```csharp
Entities\Tile.cs
<<<<<<< HEAD

=======
using Microsoft.Xna.Framework;
using Microsoft.Xna.Framework.Graphics;

namespace EpochsUnbound.Entities
{
    public class Tile
    {
        public void Update(GameTime gameTime)
        {
            // Update tile here
        }

        public void Draw(SpriteBatch spriteBatch, SpriteFont font)
        {
            // Draw tile here
        }
    }
}
>>>>>>> updated
```

Please note that this is a very basic implementation. You might need to add properties and methods to the `Tile` class depending on your game's requirements.