aider: Declared and initialized the `movementSpeed` variable in the `Game1` class constructor.
# Aider chat conversation:

USER: error with movementSpeed in game1.cs

ASSISTANT: I see that the `movementSpeed` variable is being used in the `Game1` constructor, but it's not declared anywhere in the `Game1` class. This is likely causing the error.

To fix this, we need to declare `movementSpeed` as a member of the `Game1` class.

Here is the change:

```csharp
Game1.cs
<<<<<<< HEAD
        Vector3 cameraPosition;
        Vector3 cameraDirection;
        MouseState previousMouseState;
=======
        Vector3 cameraPosition;
        Vector3 cameraDirection;
        float movementSpeed;
        MouseState previousMouseState;
>>>>>>> updated
```

And here is the change in the constructor:

```csharp
Game1.cs
<<<<<<< HEAD
            cameraPosition = new Vector3(0, 0, 0);
            cameraDirection = new Vector3(0, 0, -1);
            movementSpeed = 0.1f;

            adventurerMode = new AdventurerMode();
=======
            cameraPosition = new Vector3(0, 0, 0);
            cameraDirection = new Vector3(0, 0, -1);
            movementSpeed = 0.1f;

            adventurerMode = new AdventurerMode(movementSpeed);
>>>>>>> updated
```

Please note that I'm assuming the `AdventurerMode` constructor takes `movementSpeed` as a parameter. If it doesn't, you'll need to modify the `AdventurerMode` class to accept this parameter and use it appropriately. If you need help with that, please let me know.