aider: Added 'HexDirection' and 'Point' types in the 'Utils' namespace to represent directions in a hexagonal grid and points in a 2D space respectively.
# Aider chat conversation:

USER: 'HexDirection' error in adventurermode and 'point' errors in layout

ASSISTANT: It seems like the `HexDirection` and `Point` types are not recognized in the `AdventurerMode` and `Layout` classes respectively. This could be due to missing `using` directives for the namespaces where these types are defined.

To resolve this, we need to add the necessary `using` directives at the top of the `AdventurerMode.cs` and `Layout.cs` files. However, I don't have the `AdventurerMode.cs` file in the list of *read-write* files. Could you please add it to the chat?

For the `Layout.cs` file, I can propose an edit block to add the `using` directive for the `Point` type. But first, I need to know in which file and namespace the `Point` type is defined. Could you please provide this information?

USER: I added these *read-write* files: GameModes\AdventurerMode.cs

ASSISTANT: Great, now that we have the `AdventurerMode.cs` file, we can add the necessary `using` directive for the `HexDirection` type. However, I still need to know in which file and namespace the `HexDirection` type is defined. Could you please provide this information?

As for the `Layout.cs` file, I still need to know in which file and namespace the `Point` type is defined. Could you please provide this information as well?

USER: we need to defile ot

ASSISTANT: I see, it seems like the `HexDirection` and `Point` types are not defined yet. We need to define these types before we can use them in other files.

Let's start by defining the `HexDirection` type. This type will represent the six directions in a hexagonal grid. We can define it as an `enum` in the `Utils` namespace. Here is the *edit block* for this:

```csharp
Utils\HexDirection.cs
<<<<<<< HEAD
=======
namespace EpochsUnbound.Utils
{
    public enum HexDirection
    {
        NorthWest,
        NorthEast,
        East,
        SouthEast,
        SouthWest,
        West
    }
}
>>>>>>> updated
```

Next, let's define the `Point` type. This type will represent a point in a 2D space. We can define it as a `struct` with `x` and `y` fields in the `Utils` namespace. Here is the *edit block* for this:

```csharp
Utils\Point.cs
<<<<<<< HEAD
=======
namespace EpochsUnbound.Utils
{
    public struct Point
    {
        public double x;
        public double y;

        public Point(double x, double y)
        {
            this.x = x;
            this.y = y;
        }
    }
}
>>>>>>> updated
```

After defining these types, we need to add `using` directives for the `EpochsUnbound.Utils` namespace in the `AdventurerMode.cs` and `Layout.cs` files. I will provide the *edit blocks* for these changes once you confirm the definitions of the `HexDirection` and `Point` types.