aider: Fixed naming conflict between `EpochsUnbound.Utils.Point` and `Microsoft.Xna.Framework.Point` by using a fully qualified name for `Microsoft.Xna.Framework.Point`.
# Aider chat conversation:

USER: In Player.cs an error: "'Point' is an ambiguous reference between 'EpochsUnbound.Utils.Point' and 'Microsoft.Xna.Framework.Point'"

ASSISTANT: The error message indicates that there is a naming conflict between `EpochsUnbound.Utils.Point` and `Microsoft.Xna.Framework.Point`. Both namespaces have a `Point` class, and the compiler is unsure which one to use.

To resolve this issue, we can use a fully qualified name for the `Point` class from the `Microsoft.Xna.Framework` namespace. This will make it clear to the compiler which `Point` class we want to use.

Here is the change you need to make in the `Entities\Player.cs` file:

```csharp
Entities\Player.cs
<<<<<<< HEAD
    spriteBatch.Draw(playerTexture, new Rectangle(Position.ToPoint(), new Point(64, 64)), Color.White);
=======
    spriteBatch.Draw(playerTexture, new Rectangle(Position.ToPoint(), new Microsoft.Xna.Framework.Point(64, 64)), Color.White);
>>>>>>> updated
```

This change specifies that we want to use the `Point` class from the `Microsoft.Xna.Framework` namespace.